requirements.txt
rich
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
argparse==1.4.0
autograd==1.6.2
autograd-gamma==0.5.0
automat==22.10.0
buildtools==1.0.6
cloudpickle==3.0.0
constantly==23.10.4
contourpy==1.2.0
cycler==0.12.1
dask==2024.2.0
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.7.0
fastapi==0.109.2
fastprogress==1.0.3
filelock==3.13.1
fonttools==4.48.1
formulaic==1.0.1
fsspec==2023.12.2
furl==2.1.3
future==0.18.3
gensim==4.3.2
gitdb==4.0.11
gitpython==3.1.41
greenlet==3.0.3
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.20.3
hyperlink==21.0.0
incremental==22.10.0
interface-meta==1.3.0
joblib==1.3.2
kiwisolver==1.4.5
lifelines==0.28.0
lightning==2.1.4
lightning-utilities==0.10.1
locket==1.0.0
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
orderedmultidict==1.0.1
pandas==2.2.0
partd==1.4.1
pillow==10.2.0
pyarrow==15.0.0
pydantic==2.6.1
pydantic-core==2.16.2
pydantic-settings==2.1.0
pyparsing==3.1.1
python-dotenv==1.0.1
pytorch-lightning==2.1.4
pytz==2024.1
rdkit==2023.9.4
redo==2.0.4
regex==2023.12.25
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
selfies==2.1.1
sentry-sdk==1.40.0
setproctitle==1.3.3
simplejson==3.19.2
smart-open==6.4.0
smilespe==0.0.3
smmap==5.0.1
sniffio==1.3.0
sqlalchemy==2.0.25
starlette==0.36.3
swifter==1.4.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.1
toolz==0.12.1
torch==2.2.0
torchmetrics==1.3.0.post0
tqdm==4.66.1
transformers==4.37.2
triton==2.2.0
twisted==23.10.0
tzdata==2023.4
uvicorn==0.27.1
wandb==0.16.2
wrapt==1.16.0
zope-interface==6.1
httpx
pyjwt
python-multipart